// src/app/core/models/seafarer.model.ts
export interface Seafarer {
  Id: number;
  KinName: string | null;
  KinPhone: string | null;
  KinEmail: string | null;
  PassPortIssueDate: string | null;
  EmpId: number;
  VisaSponsorId: number;
  Remarks: string | null;
  VisaIssueDate: string | null;
  VisaExpiryDate: string | null;
  EmployeeName: string;
  SponsorName: string;
  EmployeeCode: string;
  JobName: string;
  Nationality: string;
  BirthDate: string;
  Phone: string;
  Mobile: string;
  Email: string;
  NationalId: string;
  EmploymentDate: string;
  InsuranceDate: string | null;
  PassportExpireDate: string;
  PassportNumber: string;
  Age: string;
  EmployeeNameAR: string;
  EmployeeNameEN: string;
  JobNameEN: string;
  JobNameAR: string;
  Salary: number;
  Address: string;
  Gender: string;
  Religion: string;
  InsuranceNumber: string | null;
  Area: string | null;
  MilitaryServiceExpireDate: string;
  SyndicateCardNumber: string;
  SyndicateExpirationDate: string | null;
  BusinessPhone: string | null;
  BirthPlace: string | null;
  NationalIdExpiryDate: string;
  Company: string;
  PostCategoryLevel: string;
  Tier: string;
  DrivingLicenseNumber: string;
  WorkPermit: string;
  PermitNumber: string;
  BloodType: string | null;
  WorkLocation: string | null;
  PermitExpirationDate: string;
  Notes: string | null;
  MedicalInsuranceAmount: number;
  InsuranceTitle: string | null;
  InsuranceOffice: string | null;
  MedicalInsuranceNo: string | null;
  MedicalInsuranceDate: string | null;
  GraduationYear: string;
  PostGraduateDegree: string | null;
  PostGraduateUniversity: string | null;
  PostGraduateSpecialty: string | null;
  PostGraduateGrade: string | null;
  PropationEndDate: string | null;
  EmpManager: string;
  NameOfSpouse: string | null;
  NoOfChildren: number;
  BodyWeight: number;
  Height: number;
  VisaUAEIdNO: string | null;
  NearestAirport: string;
  ResidenceNumber: string | null;
  SkypeID: string | null;
  PermanentAddressHomeCountry: string | null;
  ContactNumberHomeCountry: string | null;
  ContactNameAndNumberDuringEmergenciesUAE: string | null;
  ContactNameAndNumberDuringEmergenciesHome: string | null;
  SeamanBookNO: string | null;
  SeamanIssueDate: string | null;
  SeamanExpiryDate: string | null;
  CicpaNO: string | null;
  CicpaIssueDate: string | null;
  CicpaExpiryDate: string | null;
  Declaration: string | null;
  SignedOffFromAShipDueToMedicalReason: boolean;
  SignedOffFromAShipDueToMedicalReasonComment: string | null;
  UndergoneAnyMdicalOperation: string | null;
  UndergoneAnyMdicalOperationComment: string | null;
  DoctorConsultation: string | null;
  DoctorConsultationComment: string | null;
  HealthOrDisabilityProblem: string | null;
  HealthOrDisabilityProblemComment: string | null;
  InquiryOrInvolvedMaritimeAccident: string | null;
  InquiryOrInvolvedMaritimeAccidentComment: string | null;
  LicenseSuspendedOrRevoked: string | null;
  LicenseSuspendedOrRevokedComment: string | null;
  ConfirmationInsuranceDate: string | null;
  MaritalStatus: string;
  Status: number;
  Basic: string | null;
  Advanced: string | null;
  FullLicence: string | null;
  Maintenance: string | null;
  JoiningCost: string | null;
  PreJoiningCost: string | null;
  OnBoardingCost: string | null;
}